Probably fix for broken server auto-restart

+++ website/tunejam.py
@@ -543,26 +543,28 @@ def ChordsToHTML(chords):
     
 if __name__ == '__main__':
 
-  # Kill any old processes
-  fn = tempfile.mktemp()
-  os.system('ps aux | grep tunejam.py > %s' % fn)
-  f = open(fn)
-  lines = f.readlines()
-  f.close()
-  found_process = False
-  for line in lines:
-    pid = line.split()[1]
-    try:
-      pid = int(pid)
-      if pid != os.getpid() and not 'grep' in line:
-        print("killing pid %i" % pid)
-        os.system('kill -TERM %i' % pid)
-        found_process = True
-    except:
-      pass
-  if found_process:
-    time.sleep(1.0)
-
+  # Kill any old processes (only in outer process)
+  if 'TUNEJAM_KILLED_PROCESSES' not in os.environ:
+    os.environ['TUNEJAM_KILLED_PROCESSES'] = '1'
+    fn = tempfile.mktemp()
+    os.system('ps aux | grep tunejam.py > %s' % fn)
+    f = open(fn)
+    lines = f.readlines()
+    f.close()
+    found_process = False
+    for line in lines:
+      pid = line.split()[1]
+      try:
+        pid = int(pid)
+        if pid != os.getpid() and not 'grep' in line:
+          print("killing pid %i" % pid)
+          os.system('kill -TERM %i' % pid)
+          found_process = True
+      except:
+        pass
+    if found_process:
+      time.sleep(1.0)
+      
   # Start new server
   from os import environ
   if 'WINGDB_ACTIVE' in environ:
